+++ api/src/license/entitlements/lib/custom-permission-rules-enabled.ts
